# optimize MarshalTextTo