@peterbourgon 2024-04-07
Same as #116, please help future maintainers (and me!) understand how this works, and why it's better that the previous implementation, with some form of documentation.